+++ packages/react-components/react-components/.storybook/main.js
@@ -6,9 +6,10 @@ module.exports = /** @type {Omit<import('../../../../.storybook/main'), 'typescr
   stories: [
     ...rootMain.stories,
     '../src/**/*.stories.mdx',
-    '../src/**/*.stories.@(ts|tsx)',
+    '../src/**/index.stories.@(ts|tsx)',
     ...utils.getVnextStories(),
   ],
+  staticDirs: ['../public'],
   addons: [...rootMain.addons],
   webpackFinal: (config, options) => {
     const localConfig = { ...rootMain.webpackFinal(config, options) };