+++ packages/react-components/react-progress/src/index.ts
@@ -1,0 +1,8 @@
+export {
+  Progress,
+  progressClassNames,
+  renderProgress_unstable,
+  useProgress_unstable,
+  useProgressStyles_unstable,
+} from './Progress';
+export type { ProgressProps, ProgressSlots, ProgressState } from './Progress';