+++ packages/react-components/theme-designer/src/components/ColorTokens/ColorTokens.tsx
@@ -1,22 +1,15 @@
 /* eslint-disable react/jsx-no-bind */
 import * as React from 'react';
-import { makeStyles } from '@griffel/react';
-import { Button, Caption1 } from '@fluentui/react-components';
-import { Brands, BrandVariants, teamsDarkTheme, teamsLightTheme, Theme } from '@fluentui/react-theme';
+import { makeStyles, webDarkTheme, webLightTheme, Button, Caption1 } from '@fluentui/react-components';
+import type { Brands } from '@fluentui/react-theme';
 import { getOverridableTokenBrandColors } from './getOverridableTokenBrandColors';
-import { brandTeams } from '../../utils/brandColors';
-import type { DispatchTheme } from '../../useThemeDesignerReducer';
 import { themeNames } from '../../utils/themeList';
 import { AccessibilityList } from './AccessibilityList';
+import { AppContext } from '../../ThemeDesigner';
+import { useContextSelector } from '@fluentui/react-context-selector';
+import { brandWeb } from '../../utils/brandColors';
 
-export interface ColorTokensProps {
-  className?: string;
-  brand: BrandVariants;
-  isDark: boolean;
-  theme: Theme;
-  themeLabel: string;
-  dispatchAppState: React.Dispatch<DispatchTheme>;
-}
+export interface ColorTokensProps {}
 
 export type ColorOverrideBrands = Record<string, Brands>;
 
@@ -46,15 +39,21 @@ const useStyles = makeStyles({
   },
 });
 
-const lightBrandColors: ColorOverrideBrands = getOverridableTokenBrandColors(teamsLightTheme, brandTeams);
-const darkBrandColors: ColorOverrideBrands = getOverridableTokenBrandColors(teamsDarkTheme, brandTeams);
+const lightBrandColors: ColorOverrideBrands = getOverridableTokenBrandColors(webLightTheme, brandWeb);
+const darkBrandColors: ColorOverrideBrands = getOverridableTokenBrandColors(webDarkTheme, brandWeb);
 
 export const ColorTokens: React.FunctionComponent<ColorTokensProps> = props => {
   const styles = useStyles();
 
-  const { brand, isDark, theme, themeLabel, dispatchAppState } = props;
-
+  const appState = useContextSelector(AppContext, ctx => ctx.appState);
+  const { brand, darkOverrides, isDark, lightOverrides, theme, themeName } = appState;
+  const dispatchAppState = useContextSelector(AppContext, ctx => ctx.dispatchAppState);
   const brandColors = isDark ? darkBrandColors : lightBrandColors;
+
+  const overrides = isDark ? darkOverrides : lightOverrides;
+  const overridenTheme = { ...theme, ...overrides };
+
+  const themeLabel = themeName + (isDark ? 'Dark' : 'Light');
 
   const colorOverrideReducer: (
     state: ColorOverrides,
@@ -72,7 +71,7 @@ export const ColorTokens: React.FunctionComponent<ColorTokensProps> = props => {
       case 'Reset Overrides':
         return { ...state, [themeLabel]: {} };
       case 'Reset Custom Overrides':
-        return { ...state, customLight: {}, customDark: {} };
+        return { ...state, [themeName + 'Light']: {}, [themeName + 'Dark']: {} };
       default:
         return state;
     }
@@ -90,8 +89,21 @@ export const ColorTokens: React.FunctionComponent<ColorTokensProps> = props => {
     dispatchColorOverride({ type: 'Reset Overrides' });
   };
 
+  React.useEffect(() => {
+    // if app state overrides are reset (no entries), we want to reset local overrides if they exist
+    if (
+      (Object.keys(colorOverride[themeName + 'Light']).length > 0 ||
+        Object.keys(colorOverride[themeName + 'Dark']).length > 0) &&
+      Object.keys(appState.lightOverrides).length === 0 &&
+      Object.keys(appState.darkOverrides).length === 0
+    ) {
+      dispatchAppState({ type: 'Override' });
+      dispatchColorOverride({ type: 'Reset Custom Overrides' });
+    }
+  }, [appState]); // eslint-disable-line react-hooks/exhaustive-deps
+
   return (
-    <div className={props.className}>
+    <div className={styles.root}>
       <div className={styles.row}>
         <Caption1 className={styles.col}>Color tokens</Caption1>
         <Caption1>Assigned values</Caption1>
@@ -105,7 +117,7 @@ export const ColorTokens: React.FunctionComponent<ColorTokensProps> = props => {
         brandColors={brandColors}
         colorOverride={getCurrentOverride(themeLabel, colorOverride)}
         onNewOverride={onNewOverride}
-        theme={theme}
+        theme={overridenTheme}
       />
     </div>
   );